refactor: move session restoration logic to a trait
This commit refactors the session restoration logic into a trait `CanRestoreSession`. This improves code reusability and maintainability by consolidating the session restoration functionality into a single, easily reusable component. The original `RestoreSession` class was a static utility class, which is less flexible than a trait when needing to access instance properties.

The `RestoreSession` class has been renamed to `CanRestoreSession` and converted into a trait. The `invoke` method has been renamed to `restoreSession` and is now a protected method within the trait. This trait is used in `UpdateUserPassword` and `LogoutOtherBrowserSessionsForm` to restore the session's password hash after a password change or session logout, respectively.

The use of the trait simplifies the code in both classes and promotes a more consistent approach to session restoration throughout the application.
