+++ app/Concerns/CanRestoreSession.php
@@ -1,22 +1,27 @@
 <?php
 
-namespace App\Utils;
+namespace App\Concerns;
 
 use Illuminate\Support\Facades\Auth;
 
-class RestoreSession
+trait CanRestoreSession
 {
-    public static function invoke(): void
+    /**
+     * Restore the session's password hash after a password change or session logout.
+     */
+    protected function restoreSession(): void
     {
         $authPassword = Auth::user()?->getAuthPassword();
         $passwordHashWeb = 'password_hash_web';
-        if (request()->session()->get($passwordHashWeb, null) !== null) {
+        if (request()->hasSession() && request()->session()->get($passwordHashWeb, null)
+!== null) {
             request()->session()->put([
                 $passwordHashWeb => $authPassword,
             ]);
         }
         $passwordHashSanctum = 'password_hash_sanctum';
-        if (request()->session()->get($passwordHashSanctum, null) !== null) {
+        if (request()->hasSession() && request()->session()->get($passwordHashSanctum,
+            null) !== null) {
             request()->session()->put([
                 $passwordHashSanctum => $authPassword,
             ]);
